Move signin pages to filesystem routing
